# functions/swagger.yaml
openapi: 3.0.0
info:
  title: RMS API
  version: 1.0.0
  description: Restaurant Management System API
servers:
  - url: https://api-d4o6tbc5fq-uc.a.run.app/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /reservation/create:
    post:
      summary: Create a new reservation
      description: Create a new reservation.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tableId:
                  type: string
                startTime:
                  type: string
                  description: ISO 8601 date-time string
                endTime:
                  type: string
                  description: ISO 8601 date-time string
                people:
                  type: integer
      responses:
        '201':
          description: Reservation created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /reservation/cancel/{reservationId}:
    post:
      summary: Cancel a reservation
      description: Cancel an existing reservation.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: The ID of the reservation to cancel.
          schema:
            type: string
      responses:
        '200':
          description: Reservation cancelled successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
  /reservation/confirm/{reservationId}:
    post:
      summary: Confirm a reservation
      description: Confirm an existing reservation.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: The ID of the reservation to confirm.
          schema:
            type: string
      responses:
        '200':
          description: Reservation confirmed successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
  /reservation/complete/{reservationId}:
    post:
      summary: Complete a reservation
      description: Complete an existing reservation.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: The ID of the reservation to complete.
          schema:
            type: string
      responses:
        '200':
          description: Reservation completed successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
  /reservation/reschedule/{reservationId}:
    post:
      summary: Reschedule a reservation
      description: Reschedule an existing reservation.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: The ID of the reservation to reschedule.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                time:
                  type: string
      responses:
        '200':
          description: Reservation rescheduled successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
  /reservation/get/{reservationId}:
    get:
      summary: Get reservation details
      description: Get details of a specific reservation.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservationId
          required: true
          description: The ID of the reservation to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: Reservation details retrieved successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Reservation not found
  /reservation/user:
    get:
      summary: Get reservations for a user
      tags:
        - Reservation
      description: Get all reservations for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reservations retrieved successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /reservation/user/upcoming:
    get:
      summary: Get upcoming reservations for a user
      tags:
        - Reservation
      description: Get all upcoming reservations for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Upcoming reservations retrieved successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /reservation/free-tables:
    post:
      summary: Get free tables for a given time
      tags:
        - Reservation
      description: Get available tables for a specified time period.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startTime:
                  type: string
                  description: ISO 8601 date-time string
                endTime:
                  type: string
                  description: ISO 8601 date-time string
      responses:
        '200':
          description: Free tables retrieved successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /reservation/all:
    get:
      summary: Get all reservations
      tags:
        - Reservation
      description: Retrieve all reservations (owner only).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    tableId:
                      type: string
                    startTime:
                      type: string
                    endTime:
                      type: string
                    people:
                      type: integer
                    status:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not an owner
  /reservation/{status}:
    get:
      summary: Get reservations by status
      tags:
        - Reservation
      description: Retrieve reservations by status (owner only).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: status
          required: true
          description: The status of the reservations to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: List of reservations with the specified status
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    tableId:
                      type: string
                    startTime:
                      type: string
                    endTime:
                      type: string
                    people:
                      type: integer
                    status:
                      type: string
        '400':
          description: Invalid status
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not an owner
...
  /user:
    get:
      summary: Retrieve user information
      tags:
        - User
      description: Retrieve the details of the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
        '401':
          description: Unauthorized
  /user/all:
    get:
      summary: Retrieve all users
      tags:
        - User
      description: Retrieve a list of all users. Only accessible by owners.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
                    privileges:
                      type: string
                    phoneNumber:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...